requirements.txt
flask
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
google-generativeai
chromadb
python-dotenv